Fix bug in schema pack download
Because of the conditional imports, when doing the name change from
common to redfish_service_validator, one import got missed.  This causes
a bug where, if the schema pack directory is empty, redfish service
validator fails with an ImportError.

This commit fixes the import.

Tested:
Deleting the SchemaPack and running the service validator no longer
fails.

Signed-off-by: Ed Tanous <edtanous@google.com>